requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch=2.0.1
torchvision
transformers
diffusers
accelerate
huggingface-hub
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@stable
iopath
numpy
pillow
safetensors
pyrallis